Begin heavy refactoring
