--- SunsynkIHD/lib/fonts/fonts.h
@@ -1,7 +1,0 @@
-#ifndef _FONTS_H
-#define _FONTS_H
-
-#include "FreeSans8pt7b.h"
-#include "FreeSansBold10pt7b.h"
-
-#endif
