--- SunsynkIHD/src/SunsynkIHD35.cpp
@@ -1,288 +1,0 @@
-#include <Arduino.h>
-#include <ArduinoJson.h>
-#include <HTTPClient.h>
-#include <lvgl.h>
-#include <WiFi.h>
-#include <WiFiClientSecure.h>
-#include <WiFiMulti.h>
-#include "esp_sntp.h"
-
-#include "Config.h"
-#include "DateTime.h"
-#include "Graphics.h"
-#include "SunsynkApi.h"
-#include "Touch.h"
-
-#if defined(SCREEN_800_480)
-#include <ui_800_480.h>
-#else
-#include <ui_480_320.h>
-#endif
-
-#ifndef __CONFIG_H
-#error Configuration file is missing!
-#endif
-
-void TaskClock(void *pvParameters);
-void TaskSunsynkApi(void *pvParameters);
-void TaskOutput(void *pvParameters);
-
-TaskHandle_t TaskSunsynkApi_h;
-TaskHandle_t TaskOutput_h;
-
-// Global instance of Wi-Fi client
-WiFiMulti wiFiMulti;
-
-IhdData ihdData;
-boolean ihdDataReady = false;
-boolean ihdScreenRefreshed = false;
-unsigned long lastTouchTime = 0;
-boolean backlightOn = true;
-boolean showInfoMessage = false;
-const char* infoMessage = "";
-
-// Connect to WiFi
-void connectWifI()
-{
-    WiFi.mode(WIFI_STA);
-    wiFiMulti.addAP(WIFI_SSID, WIFI_PASSWORD);
-
-    Serial.print(" - Waiting for WiFi to connect ...");
-    while ((wiFiMulti.run() != WL_CONNECTED))
-    {
-        // Serial.print(".");
-    }
-    Serial.println(" connected");
-}
-
-void getVersion(char const *date, char const *time, char *buff)
-{
-    int month, day, year, hour, min, sec;
-    static const char month_names[] = "JanFebMarAprMayJunJulAugSepOctNovDec";
-    sscanf(date, "%s %d %d", buff, &day, &year);
-    month = (strstr(month_names, buff) - month_names) / 3 + 1;
-    year = year % 100U;
-    sscanf(time, "%d:%d:%d", &hour, &min, &sec);
-    sprintf(buff, "v%d.%02d%02d-%02d%02d", year, month, day, hour, min);
-}
-
-// Check if night mode should be enabled.
-boolean IsNightMode()
-{
-    String timeNow = getTimeString();
-
-    if (timeNow == "--:--")
-    {   
-        return false;
-    }
-
-    uint16_t timeNowInt = timeToMinutes(timeNow);
-    uint16_t timeOffInt = timeToMinutes(SCREEN_OFF_TIME);
-    uint16_t timeOnInt = timeToMinutes(SCREEN_ON_TIME);
-
-    if (lastTouchTime + SCREEN_OFF_TIMEOUT < getTime())
-    {
-        if (timeOffInt > timeOnInt)
-        {
-            if ((timeNowInt >= timeOffInt) || (timeNowInt < timeOnInt))
-            {
-                return true;
-            }
-        }
-        else
-        {
-            if ((timeNowInt >= timeOffInt) && (timeNowInt < timeOnInt))
-            {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
-
-// Enable or disable night mode.
-void SetNightMode()
-{   
-    if (IsNightMode())
-    {
-        // Get the backlight state, turn it off if it's on.
-        if (backlightOn)
-        {
-            Serial.println("Turning LCD backlight off.");
-            gfx->setBrightness(LCD_BRIGHTNESS_NIGHT);
-            backlightOn = false;
-        }
-
-        if (LCD_BRIGHTNESS_NIGHT == 0 && TaskSunsynkApi_h != NULL)
-        {
-            // Get the API task state, disable it if it's enabled
-            eTaskState state = eTaskGetState(TaskSunsynkApi_h);
-            if (state != eSuspended)
-            {
-                Serial.println("Suspending API polling task.");
-                vTaskSuspend(TaskSunsynkApi_h);
-            }
-        }
-    } else {
-        // Get the API task state, enable it if it's disabled
-        if (TaskSunsynkApi_h != NULL)
-        {
-            eTaskState state = eTaskGetState(TaskSunsynkApi_h);
-            if (state == eSuspended)
-            {
-                Serial.println("Resuming API polling task.");
-                vTaskResume(TaskSunsynkApi_h);
-            }
-        }
-
-        // Get the backlight state, turn it on if it's off
-        if (!backlightOn)
-        {
-            Serial.println("Turning LCD backlight on.");
-            gfx->setBrightness(LCD_BRIGHTNESS_DAY);
-            backlightOn = true;
-        }
-    }
-}
-
-void TaskClock(void *pvParameters)
-{
-    uint32_t time_delay = *((uint32_t *)pvParameters);
-    for (;;)
-    {
-#ifdef TFT_BL
-        SetNightMode();
-#endif
-        ihdData.time = getTimeString();
-        delay(time_delay);
-    }
-}
-
-void TaskSunsynkApi(void *pvParameters)
-{
-    uint32_t api_delay = *((uint32_t *)pvParameters);
-    for (;;)
-    {
-        GetIhdData();
-        delay(api_delay);
-    }
-}
-
-void TaskOutput(void *pvParameters)
-{
-    uint32_t output_delay = *((uint32_t *)pvParameters);
-    for (;;)
-    {
-        UpdateDisplayFields();
-        lv_timer_handler();
-        delay(output_delay);
-    }
-}
-
-void setup()
-{
-#ifdef BUZZER_PIN
-    // Turn off the buzzer
-    pinMode(BUZZER_PIN, OUTPUT);
-    ledcSetup(4, 5000, 8);
-    ledcAttachPin(BUZZER_PIN, 4);
-    ledcWrite(4, 0);
-#endif
-
-    char version[24];
-    getVersion(__DATE__, __TIME__, version);
-
-    // Initialize the serial port
-    Serial.begin(115200);
-    Serial.printf("\n%s\n", version);
-    Serial.println("\nBooting ...");
-
-    // Init Display
-    if (!gfx->begin())
-    {
-        Serial.println("gfx->begin() failed!");
-    }
-    //delay(500);
-    gfx->fillScreen(TFT_BLACK);
-
-    // Turn on the LCD backlight
-    gfx->setBrightness(LCD_BRIGHTNESS_DAY);
-
-    // Initialize LVGL
-    lv_init();
-
-    // Allocate the LVGL display buffer
-    disp_draw_buf = (lv_color_t *)heap_caps_malloc(sizeof(lv_color_t) * screenWidth * 40, MALLOC_CAP_INTERNAL | MALLOC_CAP_8BIT);
-    if (!disp_draw_buf)
-    {
-        Serial.println("LVGL disp_draw_buf allocate failed!");
-    }
-    else
-    {
-        lv_disp_draw_buf_init(&draw_buf, disp_draw_buf, NULL, screenWidth * 40);
-
-        /* Initialize the display */
-        lv_disp_drv_init(&disp_drv);
-        disp_drv.hor_res = screenWidth;
-        disp_drv.ver_res = screenHeight;
-        disp_drv.flush_cb = my_disp_flush;
-        disp_drv.draw_buf = &draw_buf;
-        lv_disp_drv_register(&disp_drv);
-
-        /* Initialize the (dummy) input device driver */
-        static lv_indev_drv_t indev_drv;
-        lv_indev_drv_init(&indev_drv);
-        indev_drv.type = LV_INDEV_TYPE_POINTER;
-        indev_drv.read_cb = my_touchpad_read;
-        lv_indev_drv_register(&indev_drv);
-    }
-
-    // Initialize the LVGL UI
-    ui_init();
-    showInfoMessage = true;
-
-    // Create the task to update the display
-    uint32_t output_delay = 10;
-    xTaskCreate(TaskOutput, "Task Output", 5120, (void *)&output_delay, 2, &TaskOutput_h);  
-
-    // Connect to Wi-Fi
-    infoMessage = "Connecting to Wi-Fi ...";
-    connectWifI();
-
-    // Connect to WiFi
-    infoMessage = "Setting the time ...";
-    configureNtpAndSetClock();
-
-    // Authenticate
-    infoMessage = "Authenticating ...";
-    GetSunsynkAuthToken();
-
-    // Create the task to get API data
-    infoMessage = "Fetching data ...";
-    uint32_t api_delay = 30000; // 30 seconds
-    xTaskCreate(TaskSunsynkApi, "Task API", 20480, (void *)&api_delay, 2, &TaskSunsynkApi_h);
-
-    // Wait up to 10s for API data
-    int delayEnd = millis() + (10 * 1000);
-    while (!ihdDataReady && millis() < delayEnd)
-    {
-        delay(100);
-    }
-
-    infoMessage = "Ready!";
-    delay(500);
-
-    // Create the task to update the time displayed on the IHD
-    uint32_t time_delay = 500; // One second
-    xTaskCreate(TaskClock, "Task Clock", 2048, (void *)&time_delay, 2, NULL);
-
-    // Hide the info panel
-    infoMessage = "";
-    showInfoMessage = false;
-}
-
-void loop()
-{
-    // Nothing to do here as all the work is done by the tasks
-}
