+++ platformio.ini
@@ -12,6 +12,7 @@
 default_envs = esp32dev
 src_dir = SunsynkIHD
 lib_dir = SunsynkIHD/lib
+include_dir = SunsynkIHD/include
 
 [env:esp32dev]
 platform = espressif32
